+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
 ## Unreleased
-- Updated SPEC (v1.0.0) for url, date and format @budhrg 
+- Updated SPEC (v1.0.0) for url, date and format @budhrg
 - Added Table of Contents for README @bexelbie
+- Fix #160: "vagrant service-manager restart openshift" not working as expected @budhrg
+- Fix #166: For CDK box, provisioners are not executed by default on Vagrant up @budhrg
 
 ## v1.0.0 Apr 07, 2016
 - Fix #132: vagrant-service-manager 1.0.0 release @navidshaikh